Code Refactor
- custom keychain and env keychain are now abstracted by an
image.Keychain function. This encapsulates the ordering of which
'keychain' is used when both flags and env variables are provided to
imgpkg
- This also made testing our mult-keychain logic that our registry
depends on much easier
- move registry and keychain into their own 'registry' package
- pass list of environment variables to the keychain (instead of calling
os.Environ itself)
- use string pre/suffix to get the authOpt instead of looping over the funcsMap
