+++ pkg/imgpkg/cmd/tag_list.go
@@ -10,6 +10,7 @@ import (
 	uitable "github.com/cppforlife/go-cli-ui/ui/table"
 	regname "github.com/google/go-containerregistry/pkg/name"
 	ctlimg "github.com/k14s/imgpkg/pkg/imgpkg/image"
+	"github.com/k14s/imgpkg/pkg/imgpkg/registry"
 	"github.com/spf13/cobra"
 )
 
@@ -21,7 +22,7 @@ type TagListOptions struct {
 	Digests       bool
 }
 
-var _ ctlimg.ImagesMetadata = ctlimg.Registry{}
+var _ ctlimg.ImagesMetadata = registry.Registry{}
 
 func NewTagListOptions(ui ui.UI) *TagListOptions {
 	return &TagListOptions{ui: ui}
@@ -41,7 +42,7 @@ func NewTagListCmd(o *TagListOptions) *cobra.Command {
 }
 
 func (o *TagListOptions) Run() error {
-	registry, err := ctlimg.NewRegistry(o.RegistryFlags.AsRegistryOpts())
+	registry, err := registry.NewRegistry(o.RegistryFlags.AsRegistryOpts())
 	if err != nil {
 		return fmt.Errorf("Unable to create a registry with the options %v: %v", o.RegistryFlags.AsRegistryOpts(), err)
 	}
